+++ .gitignore
@@ -2,3 +2,4 @@ data/
 pretrained_models/
 output/
 *.pyc
+.directory
