+++ train_rpn.py
@@ -5,7 +5,7 @@ import time
 from faster_rcnn import network
 from faster_rcnn.RPN import RPN # Hierarchical_Descriptive_Model
 from faster_rcnn.utils.timer import Timer
-from faster_rcnn.utils.HDN_utils import check_recall, checker
+from faster_rcnn.utils.net_utils import check_recall, checker
 
 from faster_rcnn.datasets.visual_genome_loader import visual_genome
 from faster_rcnn.fast_rcnn.config import cfg
